rename Layout folder to CustomLayout and change file name to index.js
