+++ src/layouts/Admin/index.js
@@ -1,4 +1,4 @@
-import CustomLayout from 'components/Layout/CustomLayout'
+import CustomLayout from 'components/CustomLayout'
 import React from 'react'
 import { withRouter } from 'react-router-dom'
