+++ src/layouts/Sensei/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { withRouter } from 'react-router-dom'
-import CustomLayout from 'components/Layout/CustomLayout'
+import CustomLayout from 'components/CustomLayout'
 
 const SenseiLayout = ({ children }) => {
   return <CustomLayout isPublic={false}>{children}</CustomLayout>
